Select the correct CCM nodelist file
The nodelist filename is determined via the PBS or SLURM jobid, and
there can be more than one in the .crayccm directory if you're running
more than one CCM job at a time.  This commit tries to select the
correct CCM nodelist file via PBS and SLURM environment variables, and
if that fails, falls back to the old method.

Note that I made the nodlist file an argument to the ccmlogin script
because once in ccmlogin under PBS the jobid changes.